flatten lib/esbuild, inline into build/prisma scripts
