--- packages/core/src/___internal-do-not-use-will-break-in-patch/load-config.ts
@@ -1,1 +1,0 @@
-export { loadConfigOnce as loadConfig } from '../lib/config';
