+++ packages/core/src/scripts/prisma.ts
@@ -1,27 +1,23 @@
+import esbuild from 'esbuild';
 import execa from 'execa';
 import { createSystem } from '../lib/createSystem';
 import {
-  getBuiltKeystoneConfigurationPath,
+  getBuiltKeystoneConfiguration,
   generateTypescriptTypesAndPrisma,
   validatePrismaAndGraphQLSchemas,
 } from '../artifacts';
-import { loadConfigOnce, loadBuiltConfig } from '../lib/config';
-import { KeystoneConfig } from './../types/config/index';
-
+import { getEsbuildConfig } from '../lib/esbuild';
 import { ExitError } from './utils';
 
 export async function prisma(cwd: string, args: string[], frozen: boolean) {
-  let config: KeystoneConfig;
   if (frozen) {
-    // TODO: this cannot be changed for now, circular dependency with getSystemPaths, getEsbuildConfig
-    const builtConfigPath = getBuiltKeystoneConfigurationPath(cwd);
-
     args = args.filter(arg => arg !== '--frozen');
-    config = loadBuiltConfig(builtConfigPath);
   } else {
-    config = await loadConfigOnce(cwd);
+    await esbuild.build(getEsbuildConfig(cwd));
   }
 
+  // TODO: this cannot be changed for now, circular dependency with getSystemPaths, getEsbuildConfig
+  const config = getBuiltKeystoneConfiguration(cwd);
   const { graphQLSchema } = createSystem(config);
   await validatePrismaAndGraphQLSchemas(cwd, config, graphQLSchema);
   await generateTypescriptTypesAndPrisma(cwd, config, graphQLSchema);
